Refactor svg_path for easier usage and less memory consumption

+++ src/svg/attributes/svg_attribute_linetypes.h
@@ -17,8 +17,8 @@ public:
   virtual bool read (const char *data, bool from_css = false) override;
   virtual bool write (QString &data, bool to_css = false) const override;
 
-  bool is_line_segment (int segment) const { return m_is_line_segment[segment]; }
-  void set_is_line_segment (int segment, bool is_line) { m_is_line_segment[segment] = is_line; }
+  bool is_line_segment (size_t segment) const { return m_is_line_segment[segment]; }
+  void set_is_line_segment (size_t segment, bool is_line) { m_is_line_segment[segment] = is_line; }
 
   void create_from_path (const svg_path *path, bool dont_create_if_exists = false);
 };
